Added a functioning code to main
The code starts SDL, loads any picture it is given (currently none), then displays a blank window. A close() function is called when the window is closed, that free loaded pictures.
Also I realized my editor puts a copyright with my name on every file, feel free to delete it.